Transpiler/pyproject.toml
[tool.poetry]
name = "transpiler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
antlr4-python3-runtime = "^4.11.1"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
pydantic = "^1.10.4"
importlib = "^1.0.4"
regex = "^2022.10.31"
black = "^22.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
